+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file.
 
+---
+
+## [1.0.1] - 2023-01-13
+
+### Features
+
+- Add findex example using dict
+- Add cover crypt example
+- Add `fetch_chains_batch_size` to findex search
+
+---
+
 ## [1.0.0] - 2022-12-21
 
 ### Documentation